Update adapter class
For clarity and ease of maintenance, adapter moved into its own source
file within it's own adapters package.
