+++ app/src/main/java/ie/wit/doityourself/activities/DIYListActivity.kt
@@ -3,17 +3,14 @@ package ie.wit.doityourself.activities
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.Menu
 import android.view.MenuItem
-import android.view.ViewGroup
+
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import ie.wit.doityourself.R
+import ie.wit.doityourself.adapters.DIYAdapter
 import ie.wit.doityourself.databinding.ActivityDiyListBinding
-import ie.wit.doityourself.databinding.CardDiytaskBinding
 import ie.wit.doityourself.main.MainApp
-import ie.wit.doityourself.models.DIYModel
 
 class DIYListActivity : AppCompatActivity() {
 
@@ -51,32 +48,5 @@ class DIYListActivity : AppCompatActivity() {
             }
         }
         return super.onOptionsItemSelected(item)
-    }
-}
-
-class DIYAdapter constructor(private var tasks: List<DIYModel>) :
-    RecyclerView.Adapter<DIYAdapter.MainHolder>() {
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MainHolder {
-        val binding = CardDiytaskBinding
-            .inflate(LayoutInflater.from(parent.context), parent, false)
-
-        return MainHolder(binding)
-    }
-
-    override fun onBindViewHolder(holder: MainHolder, position: Int) {
-        val task = tasks[holder.adapterPosition]
-        holder.bind(task)
-    }
-
-    override fun getItemCount(): Int = tasks.size
-
-    class MainHolder(private val binding : CardDiytaskBinding) :
-        RecyclerView.ViewHolder(binding.root) {
-
-        fun bind(task: DIYModel) {
-            binding.taskTitle.text = task.title
-            binding.description.text = task.description
-        }
     }
 }
